Refactor rankable talk on a mixing
